fix division by 0 in non-isotropic downsampler when computing block sizes
It used to fail in the following example: input dataset has block size [128,128,64], and pixel resolution of the data is 1,1,10. It attempted to adjust the block size in Z for the first few scale levels, but since the optimal block size is smaller than block size of the input data, it led to setting the block size to 0. At some later point (when computing the block size for the next scale level), it failed with division by 0 exception.